Refactor Copilot CLI: Introduce reasoning effort property and update session handling
