+++ extensions/copilot/src/extension/chatSessions/vscode-node/copilotCLIChatSessionInitializer.ts
@@ -16,8 +16,8 @@ import { URI } from '../../../util/vs/base/common/uri';
 import { ChatVariablesCollection, extractDebugTargetSessionIds, isPromptFile } from '../../prompt/common/chatVariablesCollection';
 import { FolderRepositoryInfo, IFolderRepositoryManager, IsolationMode } from '../common/folderRepositoryManager';
 import { emptyWorkspaceInfo, getWorkingDirectory, isIsolationEnabled, IWorkspaceInfo } from '../common/workspaceInfo';
-import { SessionIdForCLI } from '../copilotcli/common/utils';
-import { COPILOT_CLI_REASONING_EFFORT_PROPERTY, ICopilotCLIAgents, ICopilotCLIModels } from '../copilotcli/node/copilotCli';
+import { COPILOT_CLI_REASONING_EFFORT_PROPERTY, SessionIdForCLI } from '../copilotcli/common/utils';
+import { ICopilotCLIAgents, ICopilotCLIModels } from '../copilotcli/node/copilotCli';
 import { ICopilotCLISession } from '../copilotcli/node/copilotcliSession';
 import { ICopilotCLISessionService } from '../copilotcli/node/copilotcliSessionService';
 import { buildMcpServerMappings, McpServerMappings } from '../copilotcli/node/mcpHandler';
